@@ -10,7 +10,7 @@ namespace Crafting.API.Impl.Recipes
 {
     public sealed class RecipeSword : Recipe
     {
-        protected override List<IComponent> Ingredient => new List<IComponent> 
+        protected override List<IComponent> Ingredient => new()
         {
             new Handle(), new Iron(), new Iron()
         };
