@@ -5,12 +5,7 @@ namespace Crafting.API.Impl.Stats
     public sealed class Multicraft : IStat<float>
     {
         public float MAX_VALUE => 100;
-        public float Value { get; private set; }
+        public float Value { get; set; }
         public const int MAX_MULTICRAFT = 4;
-
-        public void SetValue(float newValue)
-        {
-            Value = newValue;
-        }
     }
 }
