@@ -15,17 +15,15 @@ namespace Crafting.API.Tests.BasicTests
         {
             base.Setup();
             inventory = new Inventory();
-            var inspire = sheet.GetStat<Inspiration>();
-            //inspire.SetValue(0);
         }
 
         [Test]
         public void Craft_Knowledge_Quality_Legendary_Item()
         {
             var stat = sheet.GetStat<Knowledge>();
-            stat.SetValue(stat.MAX_VALUE);
+            stat.Value = stat.MAX_VALUE;
 
-            var ouputItems = table.Craft(recipe, ingredient);
+            var ouputItems = table.Craft(recipe, Ingredient);
 
             Assert.AreEqual(Quality.Legendary, ouputItems[0].Quality);
         }
@@ -34,9 +32,9 @@ namespace Crafting.API.Tests.BasicTests
         public void Craft_Knowledge_Quality_Epic_Item()
         {
             var stat = sheet.GetStat<Knowledge>();
-            stat.SetValue(80);
+            stat.Value = 80;
 
-            var ouputItems = table.Craft(recipe, ingredient);
+            var ouputItems = table.Craft(recipe, Ingredient);
 
             Assert.AreEqual(Quality.Epic, ouputItems[0].Quality);
         }
@@ -45,9 +43,9 @@ namespace Crafting.API.Tests.BasicTests
         public void Craft_Knowledge_Quality_Rare_Item()
         {
             var stat = sheet.GetStat<Knowledge>();
-            stat.SetValue(60);
+            stat.Value = 60;
 
-            var ouputItems = table.Craft(recipe, ingredient);
+            var ouputItems = table.Craft(recipe, Ingredient);
 
             Assert.AreEqual(Quality.Rare, ouputItems[0].Quality);
         }
@@ -56,9 +54,9 @@ namespace Crafting.API.Tests.BasicTests
         public void Craft_Knowledge_Quality_Uncommon_Item()
         {
             var stat = sheet.GetStat<Knowledge>();
-            stat.SetValue(40);
+            stat.Value = 40;  
 
-            var ouputItems = table.Craft(recipe, ingredient);
+            var ouputItems = table.Craft(recipe, Ingredient);
 
             Assert.AreEqual(Quality.Uncommon, ouputItems[0].Quality);
         }
@@ -67,9 +65,9 @@ namespace Crafting.API.Tests.BasicTests
         public void Craft_Knowledge_Quality_Common_Item()
         {
             var stat = sheet.GetStat<Knowledge>();
-            stat.SetValue(0);
+            stat.Value = 0;
 
-            var ouputItems = table.Craft(recipe, ingredient);
+            var ouputItems = table.Craft(recipe, Ingredient);
 
             Assert.AreEqual(Quality.Common, ouputItems[0].Quality);
         }
@@ -77,7 +75,7 @@ namespace Crafting.API.Tests.BasicTests
         [Test]
         public void Craft_Inventory_Items_Equals()
         {
-            var ouputItems = table.Craft(recipe, ingredient);
+            var ouputItems = table.Craft(recipe, Ingredient);
 
             Assert.GreaterOrEqual(inventory.Items.Count, ouputItems.Count);
         }
