@@ -16,11 +16,11 @@ namespace Crafting.API.Tests.BasicTests
         public void Stats_Inspiration_Inspired()
         {
             var knowladge = sheet.GetStat<Knowledge>();
-            knowladge.SetValue(0);
+            knowladge.Value = 0;
             var inspire = sheet.GetStat<Inspiration>();
-            inspire.SetValue(inspire.MAX_VALUE);
+            inspire.Value = inspire.MAX_VALUE;
 
-            var ouputItem = table.Craft(recipe, ingredient);
+            var ouputItem = table.Craft(recipe, Ingredient);
 
             Assert.AreEqual(Quality.Uncommon, ouputItem[0].Quality);
         }
@@ -29,11 +29,11 @@ namespace Crafting.API.Tests.BasicTests
         public void Stats_Inspiration_Uninspired()
         {
             var knowladge = sheet.GetStat<Knowledge>();
-            knowladge.SetValue(0);
+            knowladge.Value = 0;
             var inspire = sheet.GetStat<Inspiration>();
-            inspire.SetValue(false);
+            inspire.Value = false;
 
-            var ouputItem = table.Craft(recipe, ingredient);
+            var ouputItem = table.Craft(recipe, Ingredient);
 
             Assert.AreNotEqual(Quality.Uncommon, ouputItem[0].Quality);
         }
@@ -43,9 +43,9 @@ namespace Crafting.API.Tests.BasicTests
         {
             var stat = sheet.GetStat<Multicraft>();
 
-            stat.SetValue(stat.MAX_VALUE);
+            stat.Value = stat.MAX_VALUE;
 
-            var ouputItem = table.Craft(recipe, ingredient);
+            var ouputItem = table.Craft(recipe, Ingredient);
 
             Assert.AreEqual(Multicraft.MAX_MULTICRAFT + 1, ouputItem.Count);
         }
@@ -54,11 +54,34 @@ namespace Crafting.API.Tests.BasicTests
         public void Stats_Multicraft_Minimum()
         {
             var stat = sheet.GetStat<Multicraft>();
-            stat.SetValue(0);
+            stat.Value = 0;
 
-            var ouputItem = table.Craft(recipe, ingredient);
+            var ouputItem = table.Craft(recipe, Ingredient);
 
             Assert.AreEqual(1, ouputItem.Count);
+        }
+
+        [Test]
+        public void Stats_StatSheet_SetValue_Equals()
+        {
+            var modifiedStat = sheet.SetStat<Multicraft, float>(10f);
+
+            Assert.AreEqual(modifiedStat.Value, 10f);
+        }
+
+        [Test]
+        public void Stats_StatSheet_SetValue_NotEquals()
+        {
+            try
+            {
+                var modifiedStat = sheet.SetStat<Knowledge, float>(10);
+            }
+            catch (System.Exception)
+            {
+                Assert.Pass();
+            }
+
+            Assert.Fail();
         }
     }
 }
